separating tabs in a view folder, and components for modular design and understanding
