+++ middleware.ts
@@ -1,6 +1,6 @@
-import { NextResponse, NextRequest } from "next/server";
+import { NextRequest } from "next/server";
 
 export async function middleware(request: NextRequest) {
-    const path = request.nextUrl.pathname;
-    console.log("El usuario esta intentando entrar a:", path)
+  const path = request.nextUrl.pathname;
+  console.log("El usuario esta intentando entrar a:", path);
 }
