+++ src/app/bts-fifa-2026/page.tsx
@@ -1,31 +1,15 @@
 "use client";
 
-import React, { useState, useEffect } from "react"; // importamos useCallback para el team stats
-import {
-  Music,
-  ArrowRight,
-  Disc,
-  Trophy,
-  Users,
-  BarChart3,
-} from "lucide-react";
-import { createBTSFifaUser, getBTSMatches, getBTSStats } from "@/lib/appwrite";
+import React, { Suspense } from "react"; // importamos useCallback para el team stats
+import { Trophy } from "lucide-react";
+import { Match } from "./types";
+import BTSFifa2026Content from "./components/bts-fifa-2026content";
 
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogHeader,
-  DialogTitle,
-} from "@/hooks/components/ui/dialog";
-
-import { Match, AlbumStat } from "./types";
-
-interface BracketMatchProps {
+export interface BracketMatchProps {
   allMatches: Match[];
 }
 
-const albums = [
+export const albums = [
   "2 Cool 4 Skool",
   "O!RUL8,2?",
   "Skool Luv Affair",
@@ -44,764 +28,43 @@ const albums = [
   "Arirang",
 ];
 
-const MatchSlot = ({
-  match,
-  onClick,
-}: {
-  match: Match | null;
-  onClick: () => void;
-}) => {
-  // Estilo compartido para que todo sea consistente
-  const baseClasses =
-    "relative p-1.5 sm:p-2 rounded-lg border border-white/5 bg-white/5 overflow-hidden flex justify-between items-center z-10 text-[11px] sm:text-sm w-24 sm:w-28 md:w-32 lg:w-36 xl:w-40";
-  const emptyClasses =
-    "sm:p-2 p-1.5 rounded-lg border border-white/5 bg-white/5 text-white/20 flex justify-center items-center text-[10px] sm:text-xs w-24 sm:w-28 md:w-32 lg:w-36 xl:w-40";
-
-  // RETORNO 1: Si no hay match (Pendiente)
-  if (!match) {
-    return (
-      <div
-        onClick={onClick}
-        className="match flex flex-col gap-1.5 cursor-pointer shrink-0"
-      >
-        <div className={emptyClasses}>Pending...</div>
-        <div className={emptyClasses}>Pending...</div>
-      </div>
-    );
-  }
-
-  console.log(match.team_a_start_streams)
-
-  // Lógica de cálculo (solo se ejecuta si hay match)
-  const totalMeta =
-    (match.target_streams_v2 || []).reduce(
-      (sum: number, val: number) => sum + (Number(val) || 0),
-      0,
-    ) || 1;
-
-  const getPercent = (
-    streams: number,
-    team: string, // streams lo pusimos number porque es un tipo de dato especido dentro del paritdo
-  ) =>
-    match.status === "completed" && match.winner === team
-      ? 100
-      : Math.min((Number(streams || 0) / totalMeta) * 100, 100) || 0;
-
-      console.log("Tratando de hacer opercion de porcentaje:", match.team_a_start_streams)
-
-  const percentA = getPercent(match.team_a_streams, match.team_a);
-  console.log("Porcentaje de a", percentA)
-  const percentB = getPercent(match.team_b_streams, match.team_b);
-
-  // RETORNO 2: Si hay match (Activo)
+export default function BTSFifa2026() {
   return (
-    <div
-      onClick={onClick}
-      className="match flex flex-col gap-1.5 cursor-pointer shrink-0 "
-    >
-      {/* Team A */}
-      <div
-        className={`${baseClasses} ${match.winner === match.team_a ? "bg-emerald-500/20 border-emerald-500/50" : ""}`}
-      >
-        <div
-          className={`absolute top-0 left-0 h-full bg-gradient-to-r from-purple-600/40 to-blue-500/40 -z-10 transition-all duration-300
-            ${ percentA >= 100 ? "bg-purple-700" : percentA >= 50 ? "bg-amber-500" : "bg-white"}
-           `} style={{ width: `${match.team_a_start_streams}%` }}
-        />
-        <span className="font-semibold text-white truncate flex-1 mr-1 sm:mr-2 text-[10px] sm:text-xs">
-          {match.team_a}
-        </span>
-        <span className="text-[10px] sm:text-xs text-purple-200">
-          {match.team_a_start_streams}%
-        </span>
-      </div>
-      {/* Team B */}
-      <div
-        className={`${baseClasses} ${match.winner === match.team_b ? "bg-emerald-500/20 border-emerald-500/50" : ""}`}
-      >
-        <div
-          className="absolute top-0 left-0 h-full bg-gradient-to-r from-pink-600/40 to-orange-500/40 -z-10 transition-all duration-300"
-          style={{ width: `${percentB}%` }}
-        />
-        <span className="font-semibold text-white text-left truncate flex-1 text-xs sm:text-sm">
-          {match.team_b}
-        </span>
-        <span className="text-xs text-pink-200">{Math.floor(percentB)}%</span>
-      </div>
-    </div>
-  );
-};
-
-{
-  /**NEW COMPONENT IN PROGRESS....... */
-}
-const BracketMatch2 = ({ allMatches }: BracketMatchProps) => {
-  const [selectedMatch, setSelectedMatch] = useState<Match | null>(null);
-  const [isModalOpen, setModalOpen] = useState(false);
-
-  const getWinnerOfMatch = (
-    matchA: Match | null,
-    matchB: Match | null,
-  ): Match | null => {
-    // Si alguno de los partidos de octavos no está terminado, no hay ganador para cuartos
-    if (matchA?.status !== "completed" || matchB?.status !== "completed") {
-      return null;
-    }
-
-    const winnerMatch = {
-      team_a: matchA.winner,
-      team_b: matchB.winner,
-      status: "active",
-      target_streams: matchA.target_streams_v2,
-      team_a_start_streams: 0,
-      team_b_start_stream: 0,
-
-      target_songs: [],
-      team_a_current_streams: 0,
-      team_b_current_streams: 0,
-      stage: "cuartos",
-      $id: "temp-id",
-      winner: "",
-      target_streams_v2: [],
-      team_a_streams: 0,
-      team_b_streams: 0
-    } as Match;
-    // Retornamos un objeto "partido" ficticio para los cuartos
-
-    return winnerMatch;
-  };
-
-  const handleOpen = (match: Match) => {
-    if (match) {
-      setSelectedMatch(match);
-      setModalOpen(true);
-    }
-  };
-
-  return (
-    <div className="bg-white/10 backdrop-blur-md rounded-[10px] p-4 w-full overflow-hidden lg:p-1.5">
-      <div className="w-full overflow-x-auto scrollbar-hide flex justify-start lg:justify-center cursor-grab active:cursor-grabbing">
-        {/*<div className="flex flex-row items-center justify-center gap-2 sm:gap-4 md:gap-6 lg:gap-8"> */}
-
-        <div className="flex flex-row items-center gap-2 sm:gap-4 shrink-0 min-w-max">
-          {/* Octavos Izquierda */}
-          <div className="flex flex-col gap-6 sm:gap-9">
-            {Array.from({ length: 4 }).map((_, i) => (
-              <MatchSlot
-                key={`izq-${i}`}
-                match={allMatches[i] || null}
-                onClick={() => allMatches[i] && handleOpen(allMatches[i])}
-              />
-            ))}
-          </div>
-
-          {/* Cuartos Izquierda */}
-          <div className="flex flex-col justify-around py-10 sm:py-20 gap-16 sm:gap-24">
-            <MatchSlot
-              match={getWinnerOfMatch(allMatches[0], allMatches[1])}
-              onClick={() => {}}
-            />
-
-            <MatchSlot
-              match={getWinnerOfMatch(allMatches[2], allMatches[3])}
-              onClick={() => {}}
-            />
-          </div>
-
-          {/**SEMIFINAL */}
-          <div className="flex flex-col justify-center gap-4">
-            <MatchSlot
-              match={getWinnerOfMatch(
-                getWinnerOfMatch(allMatches[0], allMatches[1]), // Ganador del Cuarto 1
-                getWinnerOfMatch(allMatches[2], allMatches[3]), // Ganador del Cuarto 2
-              )}
-              onClick={() => {}}
-            />
-          </div>
-        </div>
-
-        {/* CENTRO (FINAL) */}
-        <div className="flex flex-col items-center justify-center shrink-0 px-1 sm:px-2 md:px-4">
-          <MatchSlot
-            match={getWinnerOfMatch(
-              // Semifinal Izquierda (la que construimos antes)
-              getWinnerOfMatch(
-                getWinnerOfMatch(allMatches[0], allMatches[1]),
-                getWinnerOfMatch(allMatches[2], allMatches[3]),
-              ),
-              // Semifinal Derecha (la que construimos antes)
-              getWinnerOfMatch(
-                getWinnerOfMatch(allMatches[4], allMatches[5]),
-                getWinnerOfMatch(allMatches[6], allMatches[7]),
-              ),
-            )}
-            onClick={() => {}}
-          />
-          <div
-            className="text-xs sm:text-sm font-bold mt-2 sm:mt-4 bg-gradient-to-r from-purple-500
-           to-pink-500 bg-clip-text text-transparent"
-          >
-            Final
-          </div>
-        </div>
-
-        {/* LADO DERECHO */}
-        <div className="flex flex-row-reverse items-center gap-2 sm:gap-4 shrink-0">
-          {/* Octavos Derecha */}
-          <div className="flex flex-col gap-6 sm:gap-9">
-            {Array.from({ length: 4 }).map((_, i) => (
-              <MatchSlot
-                key={`der-${i}`}
-                match={allMatches[i + 4] || null} // Asignamos índices restantes
-                onClick={() =>
-                  allMatches[i + 4] && handleOpen(allMatches[i + 4])
-                }
-              />
-            ))}
-          </div>
-
-          {/* Cuartos Derecha */}
-          <div className="flex flex-col justify-around py-10 sm:py-20 gap-16 sm:gap-24">
-            <MatchSlot
-              match={getWinnerOfMatch(allMatches[4], allMatches[5])}
-              onClick={() => {}}
-            />
-
-            {/* Segundo slot de Cuartos Derecha: depende de Octavos 6 y 7 */}
-            <MatchSlot
-              match={getWinnerOfMatch(allMatches[6], allMatches[7])}
-              onClick={() => {}}
-            />
-          </div>
-
-          {/**SEMIFINAL */}
-          <div className="flex flex-col justify-center gap-4">
-            <MatchSlot
-              match={getWinnerOfMatch(
-                getWinnerOfMatch(allMatches[4], allMatches[5]), // Ganador del Cuarto 3
-                getWinnerOfMatch(allMatches[6], allMatches[7]), // Ganador del Cuarto 4
-              )}
-              onClick={() => {}}
-            />
-          </div>
-        </div>
-      </div>
-
-      <Dialog open={isModalOpen} onOpenChange={setModalOpen}>
-        <DialogContent className="bg-white/40">
-          <DialogHeader>
-            <DialogTitle className="text-center p-0.5">
-              🥊
-            </DialogTitle>
-            <DialogDescription asChild>
-              <div className="">
-                {selectedMatch ? ( 
-
-                  <>
-                <div className="flex flex-col text-center text-xl">
-                  <span className="font-bold"> <span className="text-purple-700 text-sm">Team A: </span>{selectedMatch.team_a}</span>
-                  <span className="bg-orange-300/55 text-center text-sm font-semibold"> VS </span>
-                  <span className="font-bold"><span className="text-purple-700 text-sm">Team B: </span>{selectedMatch.team_b}</span>
-                </div>
-
-                  <div className="bg-white/15 p-2">
-                    {selectedMatch.target_songs.map((songs, index) => (
-                      <div key={index}
-                      className="flex justify-between items-center w-full font-semibold">
-
-                        <span className="truncate"><strong>Target: </strong>{songs}</span>
-
-                        <div className="flex justify-between w-24 shrink-0">
-                          <span className=""><strong>Goal: </strong>{selectedMatch.target_streams_v2[index]}</span>
-                        </div>
-                        
-                      </div>
-                    ))}
-
-                  </div>
-                  </>
-                ) : ( 
-                  <span>No hay canciones aun....</span>
-                )
-                }
-               
-              </div>
-
-
-            </DialogDescription>
-          </DialogHeader>
-        </DialogContent>
-      </Dialog>
-
-      {/* </div> */}
-    </div>
-  );
-};
-
-{
-  /**NEW COMPONENT IN PROGRESS.... */
-}
-
-const BTSFifa2026 = () => {
-  const [formData, setFormData] = useState({
-    lastfm: "",
-    album: "",
-  });
-
-  const [isSubmitting, setIsSubmitting] = useState(false);
-  const [success, setSuccess] = useState(false);
-  const [error, setError] = useState("");
-
-  const [view, setView] = useState<"join" | "bracket" | "stats">("join");
-  const [matches, setMatches] = useState<Match[]>([]);
-  const [stats, setStats] = useState<AlbumStat[]>([]);
-  const [isLoadingMatches, setIsLoadingMatches] = useState(true);
-  const [isLoadingStats, setIsLoadingStats] = useState(true);
-
-  useEffect(() => {
-    const fetchData = async () => {
-      // Para actualizar las vistas, si no estamos en la vista correct
-      // No hacemos nada y salimos
-      if (view !== "stats" && view !== "bracket") return;
-
-      setIsLoadingMatches(true);
-      setIsLoadingStats(true);
-
-      try {
-        const [matchesData, statsData] = await Promise.all([
-          getBTSMatches(),
-          getBTSStats(),
-        ]);
-
-        if (matchesData && matchesData.length > 0) {
-          const mappedMatches : Match[] = matchesData.map((doc) => ({
-            id: doc.$id,
-            team_a: doc.team_a,
-            team_b: doc.team_b,
-            target_songs: doc.target_songs,
-            target_streams_v2: doc.target_streams_v2,
-
-            // Devolviendo lo que la interfaz pide:
-            team_a_start_streams: Number(doc.team_a_start_streams ?? 0),
-            team_b_start_stream: Number(doc.team_b_start_stream ?? doc.team_b_start_streams ?? 0),
-            team_a_current_streams: Number(doc.team_a_current_streams ||0),
-            team_b_current_streams: Number(doc.team_b_current_streams || 0),
-
-            team_a_streams:
-              Number(doc.team_a_current_streams ?? doc.team_a_start_streams ?? 0),
-            team_b_streams:
-              Number(doc.team_b_current_streams ?? doc.team_b_start_stream ?? 0),
-            stage: doc.stage,
-            status: doc.status,
-            winner: doc.winner,
-          }));
-          setMatches(mappedMatches);
-        }
-
-        if (statsData) {
-          // Sort by total_members descending
-
-          // define a narrow type for the raw documents we get from Appwrite
-          type RawStatDoc = {
-            $id: string;
-            album_name?: string;
-            total_members?: number | string;
-            status?: string;
-            [key: string]: unknown;
-          };
-
-          // keep only albums we care about (no `any`)
-          const filteredStats = (statsData as RawStatDoc[]).filter((item) =>
-            albums.includes(String(item.album_name ?? item.$id)),
-          );
-
-          // map RawStatDoc -> AlbumStat and coerce types safely
-          const mappedStats: AlbumStat[] = filteredStats.map((d) => ({
-            $id: d.$id,
-            album_name: String(d.album_name ?? d.$id ?? ""),
-            total_members:
-              typeof d.total_members === "string"
-                ? Number(d.total_members)
-                : Number(d.total_members ?? 0),
-            status: d.status ?? undefined,
-          }));
-
-          // comparator now works with guaranteed numbers
-          const sortedStats = mappedStats.sort((a, b) => b.total_members - a.total_members);
-
-          setStats(sortedStats);
-
-        }
-      } catch (err) {
-        console.error("Error fetching data:", err);
-      } finally {
-        setIsLoadingMatches(false);
-        setIsLoadingStats(false);
-      }
-    };
-
-    fetchData();
-
-    
-    // Resetemaos el feedback del usuario en view
-    setSuccess(false);
-    setError("");
-
-  }, [view]); // React "vigila" este estado
-
-  const handleChange = (
-    e: React.ChangeEvent<HTMLInputElement | HTMLSelectElement>,
-  ) => {
-    const { name, value } = e.target;
-    setFormData((prev) => ({ ...prev, [name]: value }));
-  };
-
-
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setIsSubmitting(true);
-    setError("");
-    setSuccess(false);
-
-
-    try {
-      await createBTSFifaUser(formData.lastfm, formData.album);
-
-      setSuccess(true);
-      setFormData({ lastfm: "", album: "" });
-      // Cambiar a la vista del bracket después de 2 segundos de éxito
-      setTimeout(() => {
-
-        setSuccess(false); // Agregamos para limpiar antes de navegar
-
-        setView("bracket");
-      }, 2000);
-    } catch (err : unknown ) {
-
-   const errorMessage = err instanceof Error ? err.message : String(err);
-
-      if (errorMessage.includes("ALREADY_EXISTS")) {
-        setError(
-          "¡Este usuario de Last.fm ya está registrado en un equipo! No puedes hacer trampa.",
-        );
-      } else {
-        setError("Hubo un error al unirte al equipo. Intenta nuevamente.");
-      }
-    } finally {
-      setIsSubmitting(false);
-    }
-  };
-  
-
-  return (
-    <div className="min-h-screen bg-neutral-950 p-4 lg:p-8 bg-[url('https://images.unsplash.com/photo-1518609878373-06d740f60d8b?q=80&w=2070&auto=format&fit=crop')] bg-cover bg-fixed bg-center relative font-sans">
-      {/* Overlay */}
-      <div className="absolute inset-0 bg-black/75 backdrop-blur-sm"></div>
-
-      <div className="relative z-10 max-w-7xl mx-auto">
-        {/* Header / Navigation */}
-        <div className="flex flex-col sm:flex-row justify-between items-center mb-12 border-b border-white/10 pb-6">
-          <div className="text-center sm:text-left mb-4 sm:mb-0">
-            <h1 className="text-4xl font-extrabold text-transparent bg-clip-text bg-gradient-to-r from-purple-400 via-pink-500 to-emerald-400 tracking-tight">
-              BTS x FIFA 2026
-            </h1>
-            <p className="text-neutral-400 text-sm mt-1 flex items-center justify-center sm:justify-start gap-2">
-              <Trophy className="w-4 h-4 text-yellow-500" />
-              The Ultimate Streaming Playoffs
-            </p>
-          </div>
-
-          <div className="flex bg-white/5 p-1 rounded-xl border border-white/10 overflow-x-auto whitespace-nowrap scrollbar-hide">
-            <button
-              onClick={() => setView("join")}
-              className={`px-4 sm:px-6 py-2 rounded-lg text-sm font-semibold transition-all ${view === "join" ? "bg-purple-600 text-white shadow-lg" : "text-neutral-400 hover:text-white"}`}
-            >
-              Join Team
-            </button>
-            <button
-              onClick={() => setView("bracket")}
-              className={`px-4 sm:px-6 py-2 rounded-lg text-sm font-semibold transition-all ${view === "bracket" ? "bg-pink-600 text-white shadow-lg" : "text-neutral-400 hover:text-white"}`}
-            >
-              Playoffs Bracket
-            </button>
-            <button
-              onClick={() => setView("stats")}
-              className={`px-4 sm:px-6 py-2 rounded-lg text-sm font-semibold transition-all ${view === "stats" ? "bg-blue-600 text-white shadow-lg" : "text-neutral-400 hover:text-white"}`}
-            >
-              Teams Stats
-            </button>
-          </div>
-        </div>
-
-        {/* View: Join Form */}
-        {view === "join" && (
-          <div className="flex justify-center items-center py-10 animate-in fade-in zoom-in-95 duration-500">
-            <div className="w-full max-w-md bg-white/10 backdrop-blur-xl border border-white/20 p-8 rounded-3xl shadow-2xl hover:shadow-purple-500/10 transition-all">
-              <div className="text-center mb-8">
-                <div className="w-16 h-16 bg-gradient-to-tr from-purple-500 to-pink-500 rounded-2xl mx-auto mb-4 flex items-center justify-center shadow-lg transform rotate-3">
-                  <Users className="w-8 h-8 text-white" />
-                </div>
-                <h2 className="text-2xl font-bold text-white mb-2">
-                  Choose Your Era
-                </h2>
-                <p className="text-neutral-300 text-sm">
-                  Join an album team to participate in the global streaming
-                  playoffs.
+    <Suspense
+      fallback={
+        <div className="min-h-screen bg-neutral-950 p-4 lg:p-8 bg-[url('https://images.unsplash.com/photo-1518609878373-06d740f60d8b?q=80&w=2070&auto=format&fit=crop')] bg-cover bg-fixed bg-center relative font-sans">
+          <div className="absolute inset-0 bg-black/75 backdrop-blur-sm"></div>
+          <div className="relative z-10 max-w-7xl mx-auto">
+            <div className="flex flex-col sm:flex-row justify-between items-center mb-12 border-b border-white/10 pb-6">
+              <div className="text-center sm:text-left mb-4 sm:mb-0">
+                <h1 className="text-4xl font-extrabold text-transparent bg-clip-text bg-gradient-to-r from-purple-400 via-pink-500 to-emerald-400 tracking-tight">
+                  BTS x FIFA 2026
+                </h1>
+                <p className="text-neutral-400 text-sm mt-1 flex items-center justify-center sm:justify-start gap-2">
+                  <Trophy className="w-4 h-4 text-yellow-500" />
+                  The Ultimate Streaming Playoffs
                 </p>
               </div>
-
-              <form onSubmit={handleSubmit} className="space-y-6">
-                {/* Last.fm Field */}
-                <div className="space-y-2 group">
-                  <label
-                    htmlFor="lastfm"
-                    className="text-sm font-medium text-neutral-200 flex items-center gap-2"
-                  >
-                    <Music className="w-4 h-4 text-blue-400 group-focus-within:text-blue-300 transition-colors" />
-                    Last.fm Username
-                  </label>
-                  <input
-                    type="text"
-                    id="lastfm"
-                    name="lastfm"
-                    value={formData.lastfm}
-                    onChange={(e) => {
-                      if (error) setError("");
-                      handleChange(e);
-                    }}
-                    required
-                    placeholder="e.g. musiclover99"
-                    className="w-full bg-black/50 border border-white/10 rounded-xl px-4 py-3 text-white placeholder-neutral-500 focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-transparent transition-all hover:bg-black/70"
-                  />
-                </div>
-
-                {/* Album Field */}
-                <div className="space-y-2 group">
-                  <label
-                    htmlFor="album"
-                    className="text-sm font-medium text-neutral-200 flex items-center gap-2"
-                  >
-                    <Disc className="w-4 h-4 text-emerald-400 group-focus-within:text-emerald-300 transition-colors" />
-                    Select Album (Team)
-                  </label>
-                  <div className="relative">
-                    <select
-                      id="album"
-                      name="album"
-                      value={formData.album}
-                      onChange={handleChange}
-                      required
-                      className="w-full bg-black/50 border border-white/10 rounded-xl px-4 py-3 text-white appearance-none focus:outline-none focus:ring-2 focus:ring-emerald-500 focus:border-transparent transition-all hover:bg-black/70"
-                    >
-                      <option value="" disabled className="text-neutral-500">
-                        Choose your era...
-                      </option>
-                      {albums.map((a) => (
-                        <option
-                          key={a}
-                          value={a}
-                          className="bg-neutral-900 text-white"
-                        >
-                          {a}
-                        </option>
-                      ))}
-                    </select>
-                    <div className="absolute right-4 top-1/2 -translate-y-1/2 pointer-events-none text-neutral-400">
-                      ▼
-                    </div>
-                  </div>
-                </div>
-
-                {error && (
-                  <div className="bg-red-500/20 border border-red-500/50 text-red-200 text-sm p-4 rounded-xl text-center flex items-center justify-center gap-2">
-                    <span className="w-2 h-2 rounded-full bg-red-400 animate-pulse"></span>
-                    {error}
-                  </div>
-                )}
-
-                {success && (
-                  <div className="bg-emerald-500/20 border border-emerald-500/50 text-emerald-200 text-sm p-4 rounded-xl text-center flex items-center justify-center gap-2">
-                    <span className="w-2 h-2 rounded-full bg-emerald-400 animate-pulse"></span>
-                    ¡Te has unido exitosamente! Redirigiendo a Playoffs...
-                  </div>
-                )}
-
-                {/* Submit Button */}
-                <button
-                  type="submit"
-                  disabled={isSubmitting}
-                  className="w-full bg-gradient-to-r from-purple-600 via-pink-500 to-orange-500 hover:from-purple-500 hover:via-pink-400 hover:to-orange-400 text-white font-bold py-4 px-4 rounded-xl shadow-lg transform transition-all duration-300 hover:scale-[1.02] active:scale-[0.98] disabled:opacity-70 disabled:cursor-not-allowed flex items-center justify-center gap-2 mt-6"
-                >
-                  {isSubmitting ? (
-                    <div className="w-6 h-6 border-2 border-white/30 border-t-white rounded-full animate-spin" />
-                  ) : (
-                    <>
-                      Join the Tournament
-                      <ArrowRight className="w-5 h-5" />
-                    </>
-                  )}
+              <div className="flex bg-white/5 p-1 rounded-xl border border-white/10 overflow-x-auto whitespace-nowrap scrollbar-hide">
+                <button className="px-4 sm:px-6 py-2 rounded-lg text-sm font-semibold transition-all text-neutral-400">
+                  Join Team
                 </button>
-              </form>
+                <button className="px-4 sm:px-6 py-2 rounded-lg text-sm font-semibold transition-all text-neutral-400">
+                  Playoffs Bracket
+                </button>
+                <button className="px-4 sm:px-6 py-2 rounded-lg text-sm font-semibold transition-all text-neutral-400">
+                  Teams Stats
+                </button>
+              </div>
+            </div>
+            <div className="flex justify-center items-center py-20">
+              <div className="animate-spin rounded-full h-12 w-12 border-b-2 border-purple-500"></div>
             </div>
           </div>
-        )}
-
-        {/* View: Bracket Dashboard */}
-        {view === "bracket" && (
-          <div className="animate-in fade-in slide-in-from-bottom-8 duration-700">
-            <div className="mb-8 text-center sm:text-left">
-              <h2 className="text-2xl font-bold text-white flex items-center justify-center sm:justify-start gap-3">
-                Round of 16 (Octavos de Final)
-                <span className="px-3 py-1 bg-red-500/20 text-red-400 text-xs rounded-full border border-red-500/30 flex items-center gap-2 uppercase tracking-widest">
-                  <span className="w-2 h-2 rounded-full bg-red-500 animate-pulse"></span>
-                  Live Updates
-                </span>
-              </h2>
-              <p className="text-neutral-400 mt-2 text-sm">
-                Los primeros 16 álbumes se enfrentan cara a cara. El equipo que
-                llegue primero a la meta de streams avanza a Cuartos de Final.
-              </p>
-            </div>
-
-            {/* Grid for Bracket */}
-            <div className="grid grid-cols-1 gap-6 justify-items-stretch place-items-center min-h-[200px] w-full overflow-x-auto">
-              {isLoadingMatches ? (
-                <div className="col-span-full flex flex-col items-center justify-center gap-4 text-purple-400">
-                  <div className="w-10 h-10 border-4 border-purple-500/30 border-t-purple-500 rounded-full animate-spin"></div>
-                  <p className="font-medium animate-pulse">
-                    Cargando partidos en vivo...
-                  </p>
-                </div>
-              ) : matches.length > 0 ? (
-                <BracketMatch2 allMatches={matches} />
-              ) : (
-                <div className="col-span-full text-center text-neutral-500 font-medium p-8 bg-black/30 rounded-2xl border border-white/5 w-full max-w-2xl">
-                  Aún no hay partidos activos en la base de datos.
-                </div>
-              )}
-            </div>
-
-            {/**<BracketMatch2 allMatches={partidosProcesados}/> */}
-
-            <div className="mt-12 p-6 bg-white/5 border border-white/10 rounded-2xl text-center">
-              <p className="text-neutral-400 text-sm">
-                * Las estadísticas se actualizan automáticamente según los datos
-                de Last.fm de los usuarios inscritos. Asegúrate de streamear la{" "}
-                <strong className="text-pink-400">Target Song</strong> para
-                ayudar a tu equipo.
-              </p>
-            </div>
-          </div>
-        )}
-
-        {/* View: Teams Stats */}
-        {view === "stats" && (
-          <div className="animate-in fade-in zoom-in-95 duration-500 max-w-4xl mx-auto">
-            <div className="mb-8 text-center">
-              <h2 className="text-3xl font-bold text-white flex items-center justify-center gap-3">
-                <BarChart3 className="w-8 h-8 text-blue-400" />
-                Global Teams Standings
-              </h2>
-              <p className="text-neutral-400 mt-2 text-sm">
-                Ranking oficial basado en la cantidad de ARMYs reclutados por
-                cada álbum.
-              </p>
-            </div>
-
-            <div className="bg-black/40 backdrop-blur-xl border border-white/10 rounded-3xl overflow-hidden shadow-2xl">
-              <table className="w-full text-left border-collapse">
-                <thead>
-                  <tr className="bg-white/5 text-neutral-300 text-xs uppercase tracking-widest border-b border-white/10">
-                    <th className="p-5 font-semibold text-center w-16">Rank</th>
-                    <th className="p-5 font-semibold">Album Era</th>
-                    <th className="p-5 font-semibold text-center">
-                      ARMYs Recruited
-                    </th>
-                    <th className="p-5 font-semibold text-center w-32">
-                      Status
-                    </th>
-                  </tr>
-                </thead>
-                <tbody className="text-sm">
-                  {isLoadingStats ? (
-                    <tr>
-                      <td colSpan={4} className="p-8 text-center text-blue-400">
-                        <div className="flex flex-col items-center justify-center gap-4">
-                          <div className="w-8 h-8 border-4 border-blue-500/30 border-t-blue-500 rounded-full animate-spin"></div>
-                          <p className="animate-pulse font-medium">
-                            Cargando ranking global...
-                          </p>
-                        </div>
-                      </td>
-                    </tr>
-                  ) : stats.length > 0 ? (
-                    stats.map((team, index) => (
-                      <tr
-                        key={team.$id || index}
-                        className="border-b border-white/5 hover:bg-white/5 transition-colors group"
-                      >
-                        <td className="p-5 text-center">
-                          <span
-                            className={`inline-flex items-center justify-center w-8 h-8 rounded-full font-bold ${
-                              index === 0
-                                ? "bg-yellow-500/20 text-yellow-400 border border-yellow-500/50"
-                                : index === 1
-                                  ? "bg-neutral-300/20 text-neutral-300 border border-neutral-300/50"
-                                  : index === 2
-                                    ? "bg-orange-500/20 text-orange-400 border border-orange-500/50"
-                                    : "text-neutral-500"
-                            }`}
-                          >
-                            {index + 1}
-                          </span>
-                        </td>
-                        <td className="p-5">
-                          <span className="font-bold text-white group-hover:text-blue-300 transition-colors">
-                            {team.album_name}
-                          </span>
-                        </td>
-                        <td className="p-5 text-center">
-                          <div className="flex items-center justify-center gap-2">
-                            <Users className="w-4 h-4 text-blue-400" />
-                            <span className="font-mono text-lg font-bold text-white">
-                              {team.total_members?.toLocaleString() || 0}
-                            </span>
-                          </div>
-                        </td>
-                        <td className="p-5 text-center">
-                          <span
-                            className={`px-3 py-1 rounded-full text-[10px] font-bold uppercase tracking-wider ${
-                              team.status === "active"
-                                ? "bg-emerald-500/20 text-emerald-400"
-                                : team.status === "eliminated"
-                                  ? "bg-red-500/20 text-red-400"
-                                  : "bg-neutral-500/20 text-neutral-400"
-                            }`}
-                          >
-                            {team.status || "Unknown"}
-                          </span>
-                        </td>
-                      </tr>
-                    ))
-                  ) : (
-                    <tr>
-                      <td
-                        colSpan={4}
-                        className="p-12 text-center text-neutral-500 font-medium"
-                      >
-                        Aún no hay datos de equipos en el ranking.
-                      </td>
-                    </tr>
-                  )}
-                </tbody>
-              </table>
-            </div>
-          </div>
-        )}
-      </div>
-    </div>
+        </div>
+      }
+    >
+      <BTSFifa2026Content />
+    </Suspense>
   );
-};
-
-export default BTSFifa2026;
+}
